Move data viz classes to nodes, add relative speed filtering
Removed the data visualization headers and put their implementations
directly into the data viz nodes, since including these from outside
the package is difficult anyway (maybe return to this later, unclear
if useful). Note that this broke the old "car demo" node which needs
to be revisited later anyway.

Also add filtering of targets based on relative speed,
only if car velocity is published. This may not be the best place to
do this, but then again, these data viz nodes probably shouldn't be in
this repo to begin with.