Change sign of speed in angle computation
Changed the sign of the target speed in computing the missing angle.
This depends on the convention used in translating raw data to speed
in the interface node, which still needs to be sorted out. The current
signs match positive speed towards the radar, but we wish to make
positive speed away from the radar in the future for clarity.