Fix K79 node to handle new connect message response
The K79 now responds to a connect command with the network
parameters it has stored in hardware.  Changed the message
handling to check for the correct connect response message
length, but need to interpret and debug print the network
info in the future.

Also changed a few defines to static const member variables.