Small fix to start/stop command bytes
Fixed the third byte of the start command, which specifies the number
of cycles (scans) before the radar should stop.  Setting to 0x00 means
run indefinitely. Also made 0xff "reserved" as per documentation. Tested
with K77, works well and does not stop.  Note that the number of cycles
is not currently specified in the radar documentation.