Roll back ability to filter already-rendered targets by range
Due to a runtime error encountered after changing the target visual
vectors from vectors of pointers to vectors of targets (shapes), had
to roll back the ability to filter out targets in the history buffer
based on range settings.  This was probably not being done in the
best way anyway - ideally, would want to keep all targets in history
and allow filtering without deleting.