Add enum option property to change target shape
Added ability to set shape to sphere or cube.  There are a few other
primitives (eg cone) but this is good for now.  Was hoping to use
points but this is not supported by the Shape class, and couldn't
get the point cloud-related classes to work in a custom plugin.

Still need to debug why the visualization gets laggier over time,
either a memory leak or attributed to fact that markers are not
meant for rapid update of visualization whereas point cloud is.