Refactor ainstein_rviz_plugins using new message definitions
Refactored the rviz plugin for RadarData to a plugin for RadarTargetArray
instead, removing options for raw and tracked targets.  Instead, use a
separate plugin instance for each of raw and tracked when necessary.

Compiles and works but segfaults on remove due to list property. Now that
raw and tracked are not relevant, the list property can be removed next.