Rename defines, add val gate thresh param
Renamed the noise-related KF parameter defines to STDEV to make it
clear that these are standard deviations, not variances (they get
squared in the noise matrices).

Rescoped filter parameters to a /filter namespace and added a
validation gate threshold to the filter which is set by the user based
on desired confidence level from a Chi Squared distribution with 4
DoFs (measurement dimension).  These can be looked up in a table for
now, eg the table "Lower-tail critical values of chi-square distribution
with ν degrees of freedom " on the page:

https://www.itl.nist.gov/div898/handbook/eda/section3/eda3674.htm

Also changed publisher to publish tracked targets even if list is
empty (otherwise Rviz plugin always displays last message which makes
debugging the filter difficult).  Maybe change this back later.