Add range filter for RadarTargetArray
Added a range filter to remove targets from a RadarTargetArray which
are outside specified parameter bounds for min/max range.  Added a node
and nodelet based on this filter, with the intention being that other
filters (for example, speed filter, angle filter, etc) can be run as
nodelets and stacked for easy preprocessing of raw detections.