Add collision time-based color to Rviz radar visual
Added the option to color radar target array messages based on their
time to collision (range divided by speed, corrected for direction)
in addition to normal (flat) coloring. This is similar to the Rviz
color options for PointCloud2 based on any fields in that message.

Still needs cleaning up so that the non-flat color method hides the
color selector, as well as adds configurable min/max collision times
for determining coloring. The colors are currently fixed to green at
>=10s, red at <=2s and yellow in the middle (with smooth interpolation)