Add RadarInfo RViz message filter display, compiles
Added a message filter display for RadarInfo messages in RViz. It
compiles and works fine but needs to be modified because the scaling
of the FOV cone (currently the only supported visualization for this
message type) does not work intuitively.  The position of the cone
shape visual is also in the center of the cone, so the shape must
also be translated. Come back to this later.