Fix CMake issue and move nodelets to plugins folder
Added missing dependency to range filter nodelet which was setting up a
build race condition in which the dynamic reconfigure header sometimes
was built after the nodelet, causing the build to fail.

Also created a plugins folder and moved all nodelet XML files there,
and updated the package.xml to match.