Expose radar target array to point cloud conversion
Exposed a function from ainstein_radar_filters which converts from the
RadarTargetArray message type to the custom PCL point cloud type which
includes radar data by making it static.  This was needed for new
tools which need easy access to a polar-to-cartesian function. It may
make more sense to pull out such conversions and put them in utility
class somewhere else.

Note that the only change to CMakeLists required to expose the header
from this package to any package which imports it was to add an
INCLUDE_DIRS line in the catkin_Package() function of this package.