Add latched publishing of RadarInfo msg for K79
Added a latched publication of RadarInfo message for K79 in the K79
interface class. This message is published once and stored for all
future subscribers since the radar specifications never change (as of
now).  The specifications used come from the preliminary datasheet
and are likely to change. Any unknown specifications are left at zero.