Add bounding box output from radar tracking filter
Added publication of bounding boxes for the tracked targets of the
radar target tracking filter, computed to bound all targets used for
a Kalman Filter update at each step.  This is a sort of "model-based
clustering" of radar data since the KF itself tracks with the aid of
a simple motion model.

Next, plan to add Cartesian pose+covariance output.