Add new ainstein_radar_tools subpkg
Added a new ainstein_radar_tools subpackage to ainstein_radar which is
meant to store tools and utilities based on the other subpackages but
not core to development, for example sensor fusion and SLAM nodes using
radar data among other sensors.  This could arguably be broken out into
its own package and will be if necessary, however the intent is for
these tools to aid in development for anyone using Ainstein radars.

The first and only tool in this subpackage is a simple replacement for
the "CapApp" radar/camera sensor fusion application which draws boxes
over the image to indicate targets. This requires a calibrated camera
publishing CameraInfo messages (a RealSense d435i was used for the
development). A sample launch file for using this with T79 is in the
launch folder and will probably be replaced with a ROS wiki tutorial.