Add working radar/camera fusion using TensorFlow
Added a working radar/camera fusion or "cross-validation" class which
annotates objects detected from a camera image using a pre-trained
TensorFlow-based 2d object detector with radar information for all
detected objects which overlap with radar data.

Functionality only has "runtime dependencies" on the TensorFlow object
detector in the sense that fusion is driven by radar, camera, and
detected object callbacks.  The fusion node is also prevented from
running until the object detector node advertises a service indicating
that it's ready. Finally, the label map from object index to string
name is expected to be set in the parameter server as a dictionary by
the object detector. The object detector itself could therefore be
anything which outputs vision_msgs/Detection2DArray messages,
advertises an "is ready" service and sets the label "database" (map)
in the parameter server.

For an example on how to use T79 with a RealSense d435 (RGB camera
only) and set the correct topic/service/parameter mappings, see the
launch file added in this commit.