Fix laser scan converter params, remove deprecated
Fixed the min/max range for the laserscan converted to be 0.0 and 100.0
respectively by default so that the filtering by range doesn't affect
most radars by default. These paremeters are required by the laserscan
message and should instead be set from the RadarInfo message, to be
done soon. Also removed some deprecated code.