Minor change, attempt to fix validation node output
Changed the way the radar targets are rendered on the image because the
scale is not visually useful when using actual radar parameters to set
the rectangle size. Ideally, we want the target to take up the full
screen when at the minimum range and become a point at the maximum
range. Working on this next week.