Add up to 8 synchronized topics in combine filter
Added the ability to synchronize up to 8 RadarTargetArray topics in the
combine filter. In theory the approximate time synchronizer in ROS can
support 9 topics, HOWEVER boost::bind (which is used to register the
callback) only supports 9 total arguments, the first of which in this
case must be the "this" pointer because we bind a class member function
here (see https://www.boost.org/doc/libs/1_45_0/libs/bind/bind.html#NumberOfArguments)

The ability to synchronize radar topics beyond 3 has not been tested
with data (only checked for compilation).