Add example launch for radar combine filter usage
Added a launch file example for using the radar combine filter to sync
three RadarTargetArray topics. Note that the topic names must be set
as an arg and passed in through rosparam due to the parameter being a
list of strings. Also note that the "slop" factor in message syncing
has its default set in the cfg/ file and can be loaded using the usual
dynamic reconfigure loader.