Fix race condition in KF-based tracking filter
Fixed a race condition in which the KF update loop and the radar data
callback were in a race condition, with the number of KFs possibly
changing during processing of new incoming data. Added a mutex to
synchronize these functions, however the downside is potentially
slower processing speed due to blocking waits for mutex unlocks in
the data callback. Should be fine for small numbers of KFs.