Update combine filter example launch and fix bug
Updated the radar combine filter example launch file in the
ainstein_radar_filters package to show how to launch three K79 radars
and run the combine filter.

Also fixed a bug in which the dynamic configuration was set for all
possible numbers of radars which caused runtime errors because only one
of the time sync objects was valid. Now checking the number of topics
and only setting that object.