Add support for point cloud input to radar tracking
Added the ability for the radar tracking filter to subscribe to either
a RadarTargetArray or PointCloud2 message. Internally, subscribing to
a point cloud message converts to radar message type and then calls the
original radar callback. This was added for portability since radar and
point cloud messages are used with radar sensors often interchangeably.