Add Cartesian tracking filter, needs debugging
Added a Cartesian coordinates tracking filter for radar data which
stores object state as Cartesian position and velocity rather than
spherical coordinate range/angles/speed. The advantage is that we can
track the full Cartesian velocity by using information from the radar
speed measurement (which is a projection onto the radian direction) and
consecutive range/angle measurements. This unfortunately makes the KF
into an EKF because the measurements are now nonlinear in Cartesian
space.

The filter compiles and runs but never outputs tracked targets in its
current state, and thus needs debugging. Since the RadarTarget datatype
is not capable of storing full 3d velocity anyway, we need to output
eg Pose and Twist types instead. Since the filter stores uncertainty,
we can use PoseWithCovarianceStamped and TwistWithCovarianceStamped.