Begin major tracking filter refactor to remove ROS
Refactoring the tracking filter (for now, just the spherical coordinate
filter) to be totally ROS-independent to be portable to other projects
which do not depend on ROS. Currently in the middle of refactoring.

Also, started using clang autoformatting in Visual Studio code with the
ROS plugin, so formatting will change in all files touched during this
refactor process.