Update K79 GUI update to send correct final chunk
The K79 GUI previously padded the BOOT.bin send to the device, causing
trailing zeros in the file copied to the SD card. This small change
fixes the issue, however note that the data copied to SD card was not
complete (missing the final chunk) for some reason. Still needs to be
debugged.