Add initial O-79 CAN driver, node, and launch files
Added the O-79 driver for CAN communication of raw and tracked targets
via socketcan interface, as well as nodes and corresponding launch
files. The O-79 Ethernet interface for raw/tracked targets over UDP is
exactly the same as K-79, so the K-79 node is used here for this mode
of communication. In the future, CAN and UDP output may be consolidated
into a single O-79 node.

There is a known issue with CAN data having a flipped azimuth sign, to
be fixed shortly.