Add O-79 configuration node for dynamic reconfigure
Added a node for dynamically reconfiguring the tracking filter params
stored in flash memory on the radar. Uses functionality from the new
O-79 GUI which has been added to the scripts folder (but is version
controlled in the O-79 firmware repo and will be released from there
in the future).