Update packages to build in ROS Noetic
Updated the packages to build in ROS Noetic, which required some catkin
package configuration changes as well as removing the dependency on the
jsk_perception packages. These have yet to be released in Noetic and
were only used for the BoundingBox message and RViz plugin. Instead, we
add to ainstein_radar_msgs the same bounding box/array messages and use
them instead. This prevents us from using the RViz plugin, however it
removes a major dependency. We may implement our own RViz plugin for an
aggregate message type including bounding box in the future.