Update O79 drivers with new message format
Updated the O79 drivers (CAN and UDP) to use the new message format and
message scales. The UDP target message format was modified to use two
bytes for each of range and speed, with new resolution scalers fixed to
values smaller than the smallest expected range and speed resolutions
from radar reconfiguration. The CAN message format remains the same but
uses the same new scalers. Note that these scalers are defined in both
drivers separately; they should be defined in a common header in the
future (same goes for the duplicated RadarInfo message).