Add new RViz plugin for TrackedRadarObjectArray msg
Added a new RViz plugin for the new TrackedRadarObjectArray message,
which is intended to replace the tracked RadarTargetArray message for
both spherical and Cartesian filter outputs by using a Cartesian state
for either case. The new plugin has been tested with both filters and
with elevation off and on and appears to be working well, however some
functionality still needs to be added (namely coloring based on target
ID).

The message still contains a spherical coordinates RadarTarget field
mainly to store target ID and SNR, but also has the range/azim/elev
data. This spherical data is computed from Cartesian data when using
the Cartesian filter, however this conversion does not seem to work
properly. The RadarTarget field should just be replaced with standalone
target ID and SNR fields.