Remove target from tracked object message, add ID
Removed the RadarTarget field from the RadarTrackedObject message and
replaced it with an integer object ID, since this is the only field of
the target field that was being used. This simplifies the message and
avoids converting from Cartesian to spherical which was seen to have
issues (due to frame convention).