Add range alert to Rviz plugin by scaling object
Added an "alert" to the Rviz plugin for the RadarTrackedObjectArray
message type which changes the object scale when the range is less than
the specified maximum range. The intent is to be able to easily see
when the range crosses the desired threshold; scale was used because
color mode is already used to denote different tracked objects. Note
that the range threshold for the alert can easily be extended to have a
minimum range, as well as limits on azimuth and elevation, which would
allow specifying a 3d alert region in spherical coordinates.