Add min range and min/max elevation to visual alert
Added a minimum range to the existing range alert visual indication
feature, as well as a minimum and maximum elevation (in degrees). The
min/max values are checked for consistency so that the user cannot set
eg greater min range than max range, and the user's change in this case
is overriden with the clamped value.