Remove PCL dependency from drivers pkg, WIP
Attempted to refactor the drivers package to no longer depend on PCL,
so that any install needing only the messages and drivers can run both
without having to install PCL (which is ~1GB). This commit is nearly
functional, however there is one linker error for the passthrough node
which needs to be debugged.