Update Wayv Air node using older version of new API
Updated the Wayv Air driver to use the refactored 1.1.0 version of the
new Wayv Air API (https://github.com/AinsteinAI/wayv_air_api.git branch
nick-refactor-1.1.0). This dependency has not been added explicitly (as
there appears no good way to add such a custom pip package dependency),
so the user must clone the wayv_air_api repo, check nick-refactor-1.1.0
out, and system install it per the README. The existing node has been
updated to use this API, and the old API (mmwave_CLI) has been removed.
Note that the 1.1.0 branch of the API is used due to issues with the
current version of the API outputting point clouds on linux.

The wayv_node was tested and seen to output targets (point cloud) and
tracked objects appropriately, however with lag and freezing issues to
be debugged. These may be solved when the newest API is used.