Add status message, fix CMakeLists to set python3
Added a new WayvAirStatus message to publish the board temperatures and
fixed the CMakeLists file to properly install python scripts, which has
the effect of explicitly telling ROS to use python3.

Note that the Wayv node has issues shutting down (CTRL+C does not kill
the node properly). Waiting to debug this until the 1.2.0 version of
the Wayv Air API can be used.