Add O-79 monitor alarms UDP message, not tested
Added a new message to the drivers package for O-79 "monitor alarms"
as described in the device manual. Also added a ROS-independent class
mirroring the message definition for use in the ROS-independent driver.
Both definitions use enums to enforce allowable values for the alarms.
Note that this compiles but has not been tested on hardware.