Add O-79 monitor alarms CAN message, not tested
Added a new message to the drivers package for O-79 "monitor alarms"
as described in the device manual. Analagous to the UDP driver, this
uses the MonitorAlarms ROS message type and the ROS-independent enum
class to ensure consistency. Note again that as for the UDP driver,
this compiles but has not been tested on hardware.